ml/requirements.txt
scikit-learn>=1.4
pandas>=2.2
psycopg2-binary>=2.9
joblib>=1.3
fastapi>=0.110
uvicorn>=0.29
numpy>=1.26
